Refactor schemagen tools
